pyproject.toml
[project]
name = "ll2cz"
version = "0.5.1"
description = "Transform LiteLLM database data into CloudZero AnyCost CBF format"
readme = "README.md"
authors = [
    { name = "Erik Peterson", email = "erik@cloudzero.com" }
]
license = { text = "Apache-2.0" }
keywords = ["litellm", "cloudzero", "etl", "cost-management", "billing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
requires-python = ">=3.12"
dependencies = [
    "psycopg[binary]>=3.1.0",
    "polars>=0.20.0",
    "httpx>=0.25.0",
    "connectorx>=0.3.0",
    "PyYAML>=6.0.0",
    "rich>=13.0.0",
    "litellm>=1.74.7",
]

[project.urls]
Homepage = "https://github.com/Cloudzero/cloudzero-litellm-toolkit"
Documentation = "https://github.com/Cloudzero/cloudzero-litellm-toolkit/blob/main/README.md"
Repository = "https://github.com/Cloudzero/cloudzero-litellm-toolkit.git"
Issues = "https://github.com/Cloudzero/cloudzero-litellm-toolkit/issues"
Changelog = "https://github.com/Cloudzero/cloudzero-litellm-toolkit/blob/main/CHANGELOG.md"

[project.scripts]
ll2cz = "ll2cz.cli:main"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "C"]
ignore = ["E402", "E501", "B005", "B904", "C901", "C408"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "twine>=6.1.0",
    "build>=1.2.2.post1",
]

[build-system]
requires = ["uv-build"]
build-backend = "uv_build"
